feat: Make it so that under Rachy's and Davey's ratings is a big crooked number / 100 for the average of the 3 scores they provided, make each of the inputs out of 100
The task requires modifying the input ranges for ratings to be out of 100 and displaying the average of these three ratings. The `src/routes/watermelons/+page.svelte` file is the only file that contains the rating inputs and the watermelon display logic. Therefore, all changes will be concentrated in this file. The plan addresses both parts of the request: updating the input maximums and calculating/displaying the average. The generation order is straightforward as only one file is being modified. The new `calculateAverageRating` function will be used within the same Svelte component to display the averages.